requirements.txt
autopep8==1.7.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.9.1
PyJWT==2.4.0
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
toml==0.10.2
urllib3==1.26.12
Werkzeug==2.2.2
